Making make check work
git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@7109 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ projects/RTED/RtedTransformation.cpp
@@ -108,7 +108,7 @@ void RtedTransformation::transform(SgProject* project, set<string> &rtedfiles) {
 	  && !classDecl->get_file_info()->isCompilerGenerated()
 	  ) {
 	string filename = classDecl->get_file_info()->get_filenameString();
-	int idx = filename.rfind('.');
+	size_t idx = filename.rfind('.');
 	std::string extension ="";
 	if(idx != std::string::npos)
 	  extension = filename.substr(idx+1);
